% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksj_target.R
\name{ksj_target_name}
\alias{ksj_target_name}
\title{Name of the target of a file}
\usage{
ksj_target_name(file_name)
}
\arguments{
\item{file_name}{Names of files: values of \code{file_name} in \link{ksj_available}.}
}
\value{
A character vector of target names.
}
\description{
\code{ksj_target_name()} gives the name of the target of a file in the bundled
pipeline of its dataset: \verb{N03-20250101_13_GML.zip} becomes
\code{ksj_n03_20250101_13_gml}. It is used by the pipelines and by \code{\link[=ksj_get]{ksj_get()}}.
}
\examples{
ksj_target_name("N03-20250101_13_GML.zip")
}
\keyword{internal}
