% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksjdata-package.R
\docType{package}
\name{ksjdata-package}
\alias{ksjdata}
\alias{ksjdata-package}
\title{ksjdata: National Land Numerical Information of Japan}
\description{
Provides National Land Numerical Information (Kokudo Suuchi Joho) published by the Ministry of Land, Infrastructure, Transport and Tourism of Japan. A bundled catalog lists every downloadable file with its terms of use; the selected files are downloaded and read on demand by 'targets' pipelines bundled with the package, and cached via 'tarchives'. Attributes are named after the official attribute table and product specifications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/ksjdata}
  \item \url{https://uchidamizuki.github.io/ksjdata/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/ksjdata/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

Authors:
\itemize{
  \item Mizuki Uchida \email{uchidamizuki@vivaldi.net}
}

}
\keyword{internal}
